+++ lib/src/physical/gen/battery/battery_report.dart
@@ -1,0 +1,17 @@
+library src.physical.gen.battery_report;
+
+import 'dart:io';
+
+import 'package:elec/src/physical/gen/battery/battery_optimization.dart';
+
+class BatteryReport {
+  BatteryReport({required this.optimization});
+
+  final BatteryOptimization optimization;
+
+  void dailyOperationsChartDa({required Directory directory}) {
+
+
+  }
+
+}