+++ lib/src/physical/gen/lib_battery.dart
@@ -1,2 +1,0 @@
-library src.physical.gen.lib_battery;
-