+++ reports/battery/update_mdbook.dart
@@ -1,0 +1,13 @@
+import 'package:timezone/data/latest.dart';
+
+///
+void updateIsoneChapter() {
+
+}
+
+/// Update the mdbook in rascal/presentations/energy/battery
+void main(List<String> args) {
+  initializeTimeZones();
+
+  
+}